Add check for if article actually gets scraped
As it can take some time between the OG tags for an article has been
scraped and stored in the database, to the actuall article has been
scraped and stored as a markdown file can take quite a while, and
therefore there is the possibility that an article could be logged in
the database but not scraped. This commit adds a value to the database
that will only get updated to true when the article has been scraped.