Add error catch for empty lists
Sometimes a list of ordered OG tags where one of the sources doesn't
have any article (and the list for that source is therefore empty) will
be atempted scrampled by scrampleOGTags, which will cause the program to
crash as you can't pop an element from an empty list. This commit adds
an update which will remove empty lists before it's scrambled.