Restructured how OG tags are retrivied from DB
First of all added function for for converting a list of tubles to a
list. Oftentimes SQL queries using psycopg2 return not a list of
entries, but a list of tubles, so this is hugely useful.

Also restructured how the OG tags are retrived. Instead of the function
for retrieving the OG tags also retrieving a list of profiles in the
database on its own, this list of profiles will now come from a function
paramater specified when the function is called. This allows for
retriving OG tags from only some sources depending on the user
preferences.

Of course also contains a seperate function for retrieving a list of
profiles from the database.