Fix bug with how script searches for classes
Previously, the script would be looking for classes in HTML tags, even
if the classes given in the profiles was an empty string. Instead of the
desired outcome where the script would simply return all tags regardless
of class, it would instead return all those who didn't have a class (so
those who did was excluded). This has been fixed in this update.