Use return statements instead of printing
Prior the module with the functions responsible for creating the needed
table would print to STDOUT if the table already existed. This
complicated things when it came to using the same function for giving
output to the user, and it just simply doesn't fit for a backend module
to start printing to STDOUT, so now it instead return a bool value, to
let the script that calls the function handle the output.