Add function for creating the needed users
OSINTbackend will from now on use three different accounts for
interacting with the postgresql db, osinter_admin, writer and reader.
This is done both to avoid using the default postgres user as this
should only be used with peer authentification as some systems use it
for automatic maintancne, but also to improve security, by only letting
the backend inserting, updating and selecting entries from the osinter
DB, and only letting the webserver selecting from the db, effectivly
limiting it to read access.