Split function for creating normal and superuser
Split the function that prior would create both the normal users and the
superuser. This is done because the table for storing the articles has
to be created after the superuser (since it's done using that user) but
before the normal users (to be able to give them the right priviledges
on that table).