Add some comments to extractMetaInformation and fix try except error
The extractMetaInformation function was previously build on trying to
access data from webpages in the form of ld+json data, and catching the
exception thrown if the data wasn't present (which it not always is as
the web is often quite unpredictable). The function, hovewer, didn't
catch all errors, leading to a frequent application crash. This has been
fixed now by simply catching all exceptions when trying to access the
data.