Stop adding sitename to titles in DB
Prior  the program would write the name of the newssite the article was
scraped from to the title of the article before writting it to the DB as
this looked better when rendering the HTML, but this created some
problems when looking for MD files based on the title of an article.
Therefore the program will no longer write the name of the newssite to
the title, as this can simply be rendered on the front end.