Fix order of columns when requesting OGTags from DB
Previously, the program would simply select the default order of
columnse in the article table when retrieving OGTags from the DB, which
means that if the order of the columns wasn't the default one, the
program would pair the wrong value with the wrong key. This update fixes
this by specifying the wanted columns along with the order of them.