Make createMDFile return filename and not filepath
Prior the createMDFile function would return the full path for the file
so it could be opened by obsidian, but after migrating the codebase to a
server application this isn't useful anymore. Instead the function will
now only return the filename, so it can be noted in the DB so the file
can be found again.