Add function for creating lists from OGTags
The OGTags is ordered as a dict per source, but this function will loop
through each source, pick out a single property for each time (like the
url for the article) and then put that in a list, so lists of relevant
properties is created. Is needed to use jinja templates properly,
instead of stiching together HTML manually.