Actually skip article when encountering non http 200 response
When encountering a HTTP response that wasn't 200, that intended
behavior for the script was to skip scraping that article, but instead
that would crash it. The reason for this is that when encountering non
http 200 response, scrapeWebSoup would return an empty list and instead of
skipping it, the program would try to extract OGtags from it which would
crash it. Now scrapeWebSoup will instead return None when encountering
non http 200 response, and collectOGTagsFromNewsSite will actually skip
it.