Revert commit e14c721
This commit introduced the ability to manually specify the url for the
front page of a news site to scrape urls for new articles from, but due
to an unfortunate implementation, it would prevent the program from
scraping from more than one news site. Therefore this change has been
reverted, and the problem thereby fixed now.