No password for reader and specify file perms on user basis
Previously, a bug in the code would mean that the reader user (which
should be accessible to everyone without a password) would be assigned
as password. This has been fixed.

Now, the userlist in the function used for initiate the DB users also
contains the passwordStoragePerms, which specifies which permissions the
files storing the password the respective user should be stored under,
to make it possible to have some password be readable only be the owner,
and some also readable by the group.