Force value for password hash and id for new users and simplify checkIfUserExists
From now on, when entering a new users, the id and password_hash field
will not accept a none value, as this could cause problems like not
being able to login using that user. This also means that
checkIfUserExists can be simplified as it only needs to check that a
specific username exists in the table, as this user is guarrenteed to
have both a password hash and id.